fix tests 2

+++ tests/unit/unit1323.c
@@ -48,9 +48,9 @@ static CURLcode test_unit1323(const char *arg)
       curl_mprintf("%ld.%06u to %ld.%06u got %" FMT_TIMEDIFF_T
                    ", but expected %" FMT_TIMEDIFF_T "\n",
                    (long)tests[i].first.tv_sec,
-                   tests[i].first.tv_usec,
+                   (unsigned int)tests[i].first.tv_usec,
                    (long)tests[i].second.tv_sec,
-                   tests[i].second.tv_usec,
+                   (unsigned int)tests[i].second.tv_usec,
                    result,
                    tests[i].result);
       fail("unexpected result!");
